pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-optimizer"
version = "0.1.0"
description = "A tool to optimize prompts using OpenAI API and CRISPO-inspired techniques"
authors = [{name = "Natasha Mayorga"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "ruff>=0.11.13",
]

[project.scripts]
prompt-optimizer = "prompt_optimizer.main:main"
